BREAKING!: conslidate Error logic
* Only export `CompileError` and `HttpError`
