+++ mod.ts
@@ -1,15 +1,15 @@
 // Errors
-export { HttpError } from "./src/errors/http_error.ts";
+export * as Errors from "./src/core/errors.ts";
 
 // Http
-export { Middleware } from "./src/http/middleware.ts";
-export { Request } from "./src/http/request.ts";
-export { Resource } from "./src/http/resource.ts";
-export { Response } from "./src/http/response.ts";
-export { Server } from "./src/http/server.ts";
+export { Middleware } from "./src/core/http/middleware.ts";
+export { Request } from "./src/core/http/request.ts";
+export { Resource } from "./src/core/http/resource.ts";
+export { Response } from "./src/core/http/response.ts";
+export { Server } from "./src/core/http/server.ts";
 
 // Interfaces
-export * as Interfaces from "./src/interfaces.ts";
+export * as Interfaces from "./src/core/interfaces.ts";
 
 // Types
-export * as Types from "./src/types.ts";
+export * as Types from "./src/core/types.ts";
