--- src/errors/configs_error.ts
@@ -1,17 +1,0 @@
-/**
- * Throw semantic errors related to configuration settings.
- */
-export class ConfigsError extends Error {
-  //////////////////////////////////////////////////////////////////////////////
-  // FILE MARKER - CONSTRUCTOR /////////////////////////////////////////////////
-  //////////////////////////////////////////////////////////////////////////////
-
-  /**
-   * Construct an object of this class.
-   *
-   * @param message - (optional) The exception message.
-   */
-  constructor(message: string = "Config is missing or invalid.") {
-    super(message);
-  }
-}
