--- src/errors/http_middleware_error.ts
@@ -1,9 +1,0 @@
-import { HttpError } from "./http_error.ts";
-
-/**
- * This class gives you a way to throw HTTP errors semantically in the
- * middleware classes. The difference between this class and HttpException
- * comes when you want to check which exception was thrown via
- * exception.constructor.name.
- */
-export class HttpMiddlewareError extends HttpError {}
